PythonGlue: address review concerns regarding Limited API support.
 - Remove GPL code chunk, and instead error if compiling under limited
     API with GCC < 10.
 - Make numpy build-time version requirement consistent with comment.
 - Document that certain checks in mxCreateString can be skipped under
     the limited API, as they deal with Python versions < 3.7, and the
     limited API is only used in 3.7+.
 - Move custom bdist_wheel (derived from
     https://github.com/joerick/python-abi3-package-sample) to its own
     file for licensing simplicity, and add Apache 2.0 license header
